+++ src/legacy/cxgo_client__gui__servopts__playrlst.go
@@ -195,7 +195,7 @@ func sub_4567C0(a1 *gui.Window, a2, a3, a4 uintptr) uintptr {
 						v15 = nox_xxx_objGetTeamByNetCode_418C80(int32(*((*uint32)(unsafe.Add(unsafe.Pointer(v13), 4*515)))))
 						v16 = int32(uintptr(unsafe.Pointer(v15)))
 						if v15 != nil {
-							if nox_xxx_servObjectHasTeam_419130(int32(uintptr(unsafe.Pointer(v15)))) != 0 {
+							if nox_xxx_servObjectHasTeam_419130(unsafe.Pointer(v15)) != 0 {
 								sub_4196D0(unsafe.Pointer(uintptr(v16)), unsafe.Pointer(v29), int32(*((*uint32)(unsafe.Add(unsafe.Pointer(v14), 4*515)))), 1)
 							} else {
 								nox_xxx_createAtImpl_4191D0(*(*byte)(unsafe.Add(unsafe.Pointer(v29), 57)), unsafe.Pointer(uintptr(v16)), 1, int32(*((*uint32)(unsafe.Add(unsafe.Pointer(v14), 4*515)))), 1)
